# Some low hanging performance improvements