@garethson 2018-01-28
Catching up on this issue, great work @javierhonduco.

The onnnnnly thing I might see down the road is that we start introducing conversions that aren't strictly multiplicative, like temperature units. celcius -> fahrenheit is X + (Y * Z) (probably). So you can't simply use the inverse. But this is very much future problems, and would require wholesale changes anyway.

💯