@javierhonduco 2018-01-29
@garethson That's a really good point! As you said, this is something that would require more changes, that's one of the reasons why I decided to move this PR forward, including that [we were already using the inverse conversion amount](https://github.com/Shopify/measured/blob/9eeee0c/lib/measured/unit.rb#L56)